initial implementation of background sync
